Modularize core message tests
